% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_delta}
\alias{check_delta}
\title{Check 'delta' value}
\usage{
check_delta(delta)
}
\arguments{
\item{delta}{A numeric vector between -1 and 1.}
}
\description{
Check 'delta' value
}
\keyword{internal}
