% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_flevels}
\alias{check_flevels}
\title{Check 'f' levels}
\usage{
check_flevels(f, levels)
}
\arguments{
\item{f}{A named numeric vector with non-negative values.}

\item{levels}{The groups in the data.}
}
\description{
Check 'f' levels
}
