% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_fromto}
\alias{check_fromto}
\title{Check 'from' and 'to' vectors}
\usage{
check_fromto(from, to)
}
\arguments{
\item{from}{A vector of infectors.}

\item{to}{A vector of infectees}
}
\description{
Check 'from' and 'to' vectors
}
\keyword{internal}
