% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_gamma}
\alias{check_gamma}
\title{Check 'gamma' value}
\usage{
check_gamma(gamma)
}
\arguments{
\item{gamma}{A numeric vector with non-negative values.}
}
\description{
Check 'gamma' value
}
\keyword{internal}
