% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_xn}
\alias{check_xn}
\title{Check 'x' and 'n' values}
\usage{
check_xn(x, n)
}
\arguments{
\item{x}{Number of successes.}

\item{n}{Number of trials.}
}
\description{
Check 'x' and 'n' values
}
\keyword{internal}
