% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta2gamma.R
\name{delta2gamma}
\alias{delta2gamma}
\title{Convert delta to gamma}
\usage{
delta2gamma(delta)
}
\arguments{
\item{delta}{A numeric vector of delta coefficients.}
}
\value{
A numeric vector of the corresponding gamma coefficients.
}
\description{
Convert delta to gamma
}
\examples{
delta2gamma(delta = c(-1,0,1))

}
