% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma}
\alias{gamma}
\title{Calculate Gamma Coefficient}
\usage{
gamma(pi, f)
}
\arguments{
\item{pi}{A numeric value representing the proportion of within-group transmissions.}

\item{f}{A numeric value representing the relative size of the group.}
}
\value{
The calculated Gamma value.
}
\description{
Computes the tansmission assortativity coefficient.
}
\examples{
gamma(pi = c(0.3, 0.7), f = c(0.5, 0.5))
}
