% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma2delta.r
\name{gamma2delta}
\alias{gamma2delta}
\title{Convert gamma to delta}
\usage{
gamma2delta(gamma)
}
\arguments{
\item{gamma}{A numeric vector of gamma coefficients.}
}
\value{
A numeric vector delta coefficients.
}
\description{
Convert gamma to delta
}
\examples{
gamma2delta(gamma = c(0,1,10, Inf))

}
