% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pi.R
\name{get_pi}
\alias{get_pi}
\title{Calculate Pi Values}
\usage{
get_pi(from, to, alpha = 0.05)
}
\arguments{
\item{from}{A vector of infectors.}

\item{to}{A vector of infectees.}

\item{alpha}{The significance level for the confidence intervals.}
}
\value{
A data frame with the group names and the corresponding 'pi' values and confidence intervals.
}
\description{
Calculates the within-group transmission proportion 'pi' for each group in the data.
}
\examples{
from <- c("A", "A", "NA", "C", "C", "C")
to <- c("A", "B", "B", "C", "C", "C")
get_pi(from, to, alpha = 0.05)

}
