% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{pi}
\alias{pi}
\title{Calculate Pi Values}
\usage{
pi(from, to, levels)
}
\arguments{
\item{from}{A vector of infectors.}

\item{to}{A vector of infectees.}

\item{levels}{The factor levels for the groups.}
}
\value{
A numeric vector of pi values.
}
\description{
Calculates the within-group transmission proportion
}
\examples{
from <- c("A", "A", "NA", "C", "C", "C")
to <- c("A", "B", "B", "C", "C", "C")
pi(from, to, levels = c("A", "B", "C"))

}
