% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.assortativity}
\alias{plot.assortativity}
\title{Plot Assortativity Objects}
\usage{
\method{plot}{assortativity}(x, ...)
}
\arguments{
\item{x}{An object of class "assortativity" (with subclass "gamma" or "delta").}

\item{...}{Additional arguments passed to the base \code{plot} function.}
}
\description{
Plots error bars for objects of class "assortativity" (with subclasses "gamma" and "delta").
The plot includes a horizontal dotted line at y = 1 for subclass "gamma" and y = 0 for subclass "delta", referring to the neutral value.
}
\examples{
from <- c("A", "A", NA, "C", "C", "C")
to <- c("A", "B", "B", "C", "C", "C")
f <- c(A = 0.3, B = 0.2, C = 0.5)
plot(get_delta(from, to, f))
plot(get_gamma(from, to, f))

}
