% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma2delta.r
\name{gamma2delta}
\alias{gamma2delta}
\title{Convert Gamma to Delta}
\usage{
gamma2delta(gamma)
}
\arguments{
\item{gamma}{A numeric value representing the gamma parameter.}
}
\value{
A numeric value representing the delta parameter.
}
\description{
Converts a gamma value to a delta value.
}
\examples{
gamma2delta(gamma = c(0,1,10, Inf))

}
